# docs/manager-api-openapi.yaml
openapi: 3.0.3
info:
  title: CrackHash Manager API
  description: |
    API менеджера системы CrackHash.

    Менеджер предоставляет External API для клиентов, отправляющих запросы на взлом хэша.
    Взаимодействие с воркерами происходит через RabbitMQ очереди (tasks, tasks_progress, dead_letter).
  version: 2.0.0
  contact:
    name: CrackHash Team

servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://manager-service:8080
    description: Manager service in Docker network

tags:
  - name: External API
    description: API для клиентов - создание задач и получение результатов
  - name: Health
    description: Проверка работоспособности сервиса

paths:
  /api/hash/crack:
    post:
      summary: Создать запрос на взлом хэша
      description: |
        Принимает MD5-хэш и максимальную длину искомого слова.
        Создаёт подзадачи, сохраняет их в MongoDB со статусом "pending".
        Фоновая горутина отправляет подзадачи в RabbitMQ очередь "tasks" для обработки воркерами.
        Возвращает идентификатор запроса для отслеживания статуса.
      operationId: createCrackTask
      tags:
        - External API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrackRequest'
            examples:
              example1:
                summary: Взлом хэша слова "abcd"
                value:
                  hash: "e2fc714c4727ee9395f324cd2e7f331f"
                  maxLength: 4
              example2:
                summary: Взлом хэша слова "test"
                value:
                  hash: "098f6bcd4621d373cade4e832627b4f6"
                  maxLength: 4
      responses:
        '200':
          description: Запрос успешно создан и сохранён в БД (с репликацией)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrackResponse'
        '400':
          description: Неверный формат запроса
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string

  /api/hash/status:
    get:
      summary: Получить статус выполнения запроса
      description: |
        Возвращает текущий статус выполнения запроса на взлом хэша.
        Менеджер агрегирует прогресс от всех воркеров из MongoDB.

        Возможные статусы:
        - `NOT_STARTED` - задача создана, подзадачи ещё не обработаны
        - `IN_PROGRESS` - задача выполняется воркерами
        - `READY` - все подзадачи завершены успешно
        - `ERROR` - подзадача попала в DLQ (например, стоп-слово "bom")
      operationId: getTaskStatus
      tags:
        - External API
      parameters:
        - name: requestId
          in: query
          required: true
          description: UUID запроса, полученный при создании
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Статус запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
              examples:
                not_started:
                  summary: Задача не начата
                  value:
                    status: "NOT_STARTED"
                    progress: 0
                    data: []
                in_progress:
                  summary: Задача выполняется
                  value:
                    status: "IN_PROGRESS"
                    progress: 65
                    data: []
                ready:
                  summary: Задача завершена успешно
                  value:
                    status: "READY"
                    progress: 100
                    data: ["abcd"]
                error:
                  summary: Подзадача в DLQ (стоп-слово)
                  value:
                    status: "ERROR"
                    progress: 50
                    data: []
        '400':
          description: Неверный формат requestId
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string

  /health:
    get:
      summary: Health check endpoint
      description: Проверка работоспособности менеджера
      operationId: healthCheck
      tags:
        - Health
      responses:
        '200':
          description: Менеджер работает нормально
          content:
            text/plain:
              schema:
                type: string
              example: "OK"

components:
  schemas:
    CrackRequest:
      type: object
      required:
        - hash
        - maxLength
      properties:
        hash:
          type: string
          description: MD5 хэш слова, которое нужно найти
          pattern: '^[a-f0-9]{32}$'
          example: "e2fc714c4727ee9395f324cd2e7f331f"
        maxLength:
          type: integer
          description: Максимальная длина искомого слова (от 1 до maxLength включительно). Также определяет количество подзадач.
          minimum: 1
          maximum: 10
          example: 4

    CrackResponse:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
          format: uuid
          description: Уникальный идентификатор запроса для отслеживания статуса
          example: "550e8400-e29b-41d4-a716-446655440000"

    TaskStatus:
      type: object
      required:
        - status
        - progress
        - data
      properties:
        status:
          type: string
          enum:
            - NOT_STARTED
            - IN_PROGRESS
            - READY
            - ERROR
          description: |
            Текущий статус выполнения запроса:
            * `NOT_STARTED` - подзадачи ещё не обработаны ни одним воркером
            * `IN_PROGRESS` - воркеры обрабатывают подзадачи
            * `READY` - все подзадачи завершены
            * `ERROR` - хотя бы одна подзадача попала в DLQ
        progress:
          type: integer
          description: Процент выполнения (0-100)
          minimum: 0
          maximum: 100
        data:
          type: array
          items:
            type: string
          description: Найденные слова, соответствующие хэшу